requirements.txt
# Live Talker - Requirements

# Core dependencies
numpy>=1.21.0
pyaudio>=0.2.11
soundfile>=0.11.0
scipy>=1.7.0

# ASR Engines
faster-whisper>=0.9.0          # Whisper ASR
funasr>=0.8.0                  # FunASR (Chinese-optimized)
modelscope>=1.9.0              # FunASR dependency
sherpa-onnx>=1.12.0            # FireRedASR (optional)

# TTS Engines
edge-tts>=6.1.0                # Edge-TTS (free, high quality)
pyttsx3>=2.90                  # Pyttsx3 (offline)

# VAD
torch>=2.0.0                   # For Silero VAD
torchaudio>=2.0.0              # Required by FunASR
webrtcvad>=2.0.10              # WebRTC VAD (optional, requires C++ compiler)

# LLM
openai>=1.0.0                  # Deepseek API (compatible with OpenAI)
requests>=2.28.0               # HTTP requests

# Audio processing
librosa>=0.9.2                 # Audio analysis
pydub>=0.25.1                  # Audio format conversion

# Utilities
python-dotenv>=0.19.0          # Environment variables
pyyaml>=6.0                    # YAML config

# Logging
colorlog>=6.7.0                # Colored logging

